Refactored code, added new controllers, interfaces and configuration with db

+++ LES.Domain/Models/Product.cs
@@ -4,10 +4,10 @@
     {
         public Guid Id { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Qtd { get; set; }
         public bool IsActive { get; set; }
         public Guid CategoryId { get; set; }
         public virtual Category Category { get; set; }
-    }
+	}
 }
